Use video URI as URL again and fix resolution fallback
Since using the new endpoint for video URLs, videos were not
marked as watched anymore. This is because the video URLs were
now containing the direct media links. But they are not unique,
because they contain information that changes on every request.
By using video URI as URL again, we get unique URLs again.

Also fixed a bug were the resolution-not-found fallback
would fall back to AV1 codec instead of H264.

Fixes #25