Overwrite macOS user-agent when requesting videos
Kodi automatically uses a operating system based User-Agent
for HTTP requests. This causes an issue with the Vimeo API
endpoint when a request is made from macOS, because the
Vimeo endpoint then thinks it can deliver the DASH format
to an iOS device. This DASH format can currently not be
played by ffmpeg. By appending a custom User-Agent at the
end of the URL, the User-Agent can be overwritten.

Fixes #31