Don't add video to playlist when shared
When a video is shared with this addon, it can be
done by calling `Playlist.Add` or `Player.Open`.
If it is done via `Playlist.Add`, the item is already
in a playlist, so we don't have to do it ourselves.
If it is done via `Player.Open`, the user probably
doesn't want to create a playlist at all.

Resolves #66