Handle geo-blocking errors
Vimeo decided to geo-block API endpoints dedicated
to search, channels and a lot of other things. The
only thing still working are video details. So the
only useful thing you can do with this addon is to
share Vimeo videos with Kodi.

To make the user understand that Vimeo is blocking
requests and the addon is not broken, a dedicated
error notification was added.